(рефакторинг) Выделил светофор в отдельный модуль
